fix: move `stage2.ts` to `src`
