+++ src/stage2.ts
